Added code to simplify knowing where the error is.
Create a list of tests, and using map, run each.
This is the basic of testing automatically. Each of the functions can return true or false depending
on whether they pass the test or not. run_tests would then count how many functions pass and how many
functions fail.

I will leave that for all of you to improve on, if you are interested.